+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 6.6.4 (8/9/2018 - )
 * kafka: update to 2.0.0
+* redis: check values must not be empty in encoding
 
 ### 6.6.3 (8/9/2018 - 8/9/2018)
 * redis: support List